+++ test/system/test_seq_copy_sys.py
@@ -56,7 +56,7 @@ _SEED_DEV = 17
      " --rnn-attention-type mhdot --rnn-attention-num-hidden 32 --batch-size 16 --rnn-attention-mhdot-heads 1"
      " --loss cross-entropy --optimized-metric perplexity --max-updates 6000"
      " --transformer-attention-heads 4 --transformer-model-size 32"
-     " --transformer-feed-forward-num-hidden 64"
+     " --transformer-feed-forward-num-hidden 64 --transformer-activation-type gelu"
      " --checkpoint-frequency 1000 --optimizer adam --initial-learning-rate 0.001",
      "--beam-size 5",
      1.01,
@@ -132,7 +132,7 @@ def test_seq_copy(name, train_params, translate_params, perplexity_thresh, bleu_
      " --rnn-attention-type mhdot --rnn-attention-num-hidden 32 --batch-size 16 --rnn-attention-mhdot-heads 2"
      " --loss cross-entropy --optimized-metric perplexity --max-updates 5000"
      " --transformer-attention-heads 4 --transformer-model-size 32"
-     " --transformer-feed-forward-num-hidden 64"
+     " --transformer-feed-forward-num-hidden 64 --transformer-activation-type gelu"
      " --checkpoint-frequency 1000 --optimizer adam --initial-learning-rate 0.001",
      "--beam-size 5",
      1.02,